Completed the Academic version.
The academic version is working. It has been tested, and it is working
properly. There was an addition to the engine to calculate the optimal
solution. The Academic Handler was also edited to call the new engine
method to retrieve the optimal solution and the amount of moves to
display proper message when the game is finished.